# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal AI Development Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Add a project (local folder or GitHub clone)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Remove a project (does not delete files)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /projects/{id}/open:
    post:
      operationId: openProject
      tags: [projects]
      summary: Set project as active
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── FILES ─────────────────────────────────────────────────────────────────
...
  /secrets:
    get:
      operationId: listSecrets
      tags: [secrets]
      summary: List secret names (never values) for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Secret names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecretEntry"

    post:
      operationId: createSecret
      tags: [secrets]
      summary: Add a secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /secrets/{id}:
    patch:
      operationId: updateSecret
      tags: [secrets]
      summary: Update a secret value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteSecret
      tags: [secrets]
      summary: Delete a secret
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── CHECKPOINTS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Project:
      type: object
      required: [id, name, localPath, status, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        localPath:
          type: string
        githubUrl:
          type: ["string", "null"]
        activeBranch:
          type: ["string", "null"]
        previewPort:
          type: ["integer", "null"]
        startCommand:
          type: ["string", "null"]
        buildCommand:
          type: ["string", "null"]
        status:
          type: string
          enum: [running, stopped, error, idle]
        isActive:
          type: boolean
        lastOpenedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        localPath:
          type: string
        githubUrl:
          type: string
        startCommand:
          type: string
        buildCommand:
          type: string
        previewPort:
          type: number
    ProjectPatch:
      type: object
      properties:
        name:
          type: string
        startCommand:
          type: string
        buildCommand:
          type: string
        previewPort:
          type: number
        activeBranch:
          type: string
...
    SecretEntry:
      type: object
      required: [id, projectId, name, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        name:
          type: string
        createdAt:
          type: string
    SecretInput:
      type: object
      required: [projectId, name, value]
      properties:
        projectId:
          type: number
        name:
          type: string
          minLength: 1
        value:
          type: string
    SecretPatch:
      type: object
      required: [value]
      properties:
        value:
          type: string

    # Checkpoints
...